% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pray.R
\name{shinto}
\alias{shinto}
\title{Display Shinto Shrine ASCII Art}
\usage{
shinto()
}
\description{
The \code{shinto()} function outputs a simple ASCII art representation of a Shinto shrine to the console.
}
\examples{
shinto()
}
